Reorganize shared models
