+++ client/src/app/+videos/video-list/overview/overview.service.ts
@@ -5,7 +5,8 @@ import { Injectable } from '@angular/core'
 import { RestExtractor, ServerService } from '@app/core'
 import { immutableAssign } from '@app/helpers'
 import { VideoService } from '@app/shared/shared-main'
-import { peertubeTranslate, VideosOverview as VideosOverviewServer } from '@shared/models'
+import { peertubeTranslate } from '@shared/core-utils/i18n'
+import { VideosOverview as VideosOverviewServer } from '@shared/models'
 import { environment } from '../../../../environments/environment'
 import { VideosOverview } from './videos-overview.model'
