+++ client/src/app/app.component.ts
@@ -13,7 +13,8 @@ import { InstanceConfigWarningModalComponent } from '@app/modal/instance-config-
 import { WelcomeModalComponent } from '@app/modal/welcome-modal.component'
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap'
 import { I18n } from '@ngx-translate/i18n-polyfill'
-import { BroadcastMessageLevel, getShortLocale, is18nPath, ServerConfig, UserRole } from '@shared/models'
+import { getShortLocale, is18nPath } from '@shared/core-utils/i18n'
+import { BroadcastMessageLevel, ServerConfig, UserRole } from '@shared/models'
 import { MenuService } from './core/menu/menu.service'
 import { POP_STATE_MODAL_DISMISS } from './helpers'
 import { peertubeLocalStorage } from '@root-helpers/peertube-web-storage'
