+++ client/src/app/app.module.ts
@@ -5,7 +5,7 @@ import { BrowserModule } from '@angular/platform-browser'
 import { ServerService } from '@app/core'
 import localeOc from '@app/helpers/locales/oc'
 import { MetaLoader, MetaModule, MetaStaticLoader, PageTitlePositioning } from '@ngx-meta/core'
-import { buildFileLocale, getCompleteLocale, isDefaultLocale } from '@shared/models'
+import { buildFileLocale, getCompleteLocale, isDefaultLocale } from '@shared/core-utils/i18n'
 import { AppRoutingModule } from './app-routing.module'
 import { AppComponent } from './app.component'
 import { CoreModule } from './core'
