+++ client/src/app/core/server/server.service.ts
@@ -4,15 +4,8 @@ import { HttpClient } from '@angular/common/http'
 import { Inject, Injectable, LOCALE_ID } from '@angular/core'
 import { getDevLocale, isOnDevLocale, sortBy } from '@app/helpers'
 import { peertubeLocalStorage } from '@root-helpers/peertube-web-storage'
-import {
-  getCompleteLocale,
-  isDefaultLocale,
-  peertubeTranslate,
-  SearchTargetType,
-  ServerConfig,
-  ServerStats,
-  VideoConstant
-} from '@shared/models'
+import { getCompleteLocale, isDefaultLocale, peertubeTranslate } from '@shared/core-utils/i18n'
+import { SearchTargetType, ServerConfig, ServerStats, VideoConstant } from '@shared/models'
 import { environment } from '../../../environments/environment'
 
 @Injectable()
