+++ client/src/assets/player/peertube-player-manager.ts
@@ -19,7 +19,8 @@ import './videojs-components/settings-panel'
 import './videojs-components/settings-panel-child'
 import './videojs-components/theater-button'
 import videojs from 'video.js'
-import { isDefaultLocale, VideoFile } from '@shared/models'
+import { VideoFile } from '@shared/models'
+import { isDefaultLocale } from '@shared/core-utils/i18n'
 import { RedundancyUrlManager } from './p2p-media-loader/redundancy-url-manager'
 import { segmentUrlBuilderFactory } from './p2p-media-loader/segment-url-builder'
 import { segmentValidatorFactory } from './p2p-media-loader/segment-validator'
