+++ client/webpack/webpack.video-embed.js
@@ -29,6 +29,7 @@ module.exports = function () {
         'video.js$': path.resolve('node_modules/video.js/core.js'),
         '@root-helpers': path.resolve('src/root-helpers'),
         '@shared/models': path.resolve('../shared/models'),
+        '@shared/core-utils': path.resolve('../shared/core-utils')
       }
     },
