+++ shared/models/feeds/feed-format.enum.ts
@@ -1,4 +1,4 @@
-export enum FeedFormat {
+export const enum FeedFormat {
   RSS = 'xml',
   ATOM = 'atom',
   JSON = 'json'
