+++ shared/models/index.ts
@@ -7,7 +7,6 @@ export * from './redundancy'
 export * from './users'
 export * from './videos'
 export * from './feeds'
-export * from './i18n'
 export * from './overviews'
 export * from './plugins'
 export * from './search'
