+++ shared/models/moderation/abuse/abuse-reason.model.ts
@@ -1,4 +1,4 @@
-export enum AbusePredefinedReasons {
+export const enum AbusePredefinedReasons {
   VIOLENT_OR_REPULSIVE = 1,
   HATEFUL_OR_ABUSIVE,
   SPAM_OR_MISLEADING,
@@ -18,16 +18,3 @@ export type AbusePredefinedReasonsString =
   'serverRules' |
   'thumbnails' |
   'captions'
-
-export const abusePredefinedReasonsMap: {
-  [key in AbusePredefinedReasonsString]: AbusePredefinedReasons
-} = {
-  violentOrRepulsive: AbusePredefinedReasons.VIOLENT_OR_REPULSIVE,
-  hatefulOrAbusive: AbusePredefinedReasons.HATEFUL_OR_ABUSIVE,
-  spamOrMisleading: AbusePredefinedReasons.SPAM_OR_MISLEADING,
-  privacy: AbusePredefinedReasons.PRIVACY,
-  rights: AbusePredefinedReasons.RIGHTS,
-  serverRules: AbusePredefinedReasons.SERVER_RULES,
-  thumbnails: AbusePredefinedReasons.THUMBNAILS,
-  captions: AbusePredefinedReasons.CAPTIONS
-}
