+++ shared/models/moderation/abuse/abuse-state.model.ts
@@ -1,4 +1,4 @@
-export enum AbuseState {
+export const enum AbuseState {
   PENDING = 1,
   REJECTED = 2,
   ACCEPTED = 3
