+++ shared/models/plugins/hook-type.enum.ts
@@ -1,4 +1,4 @@
-export enum HookType {
+export const enum HookType {
   STATIC = 1,
   ACTION = 2,
   FILTER = 3
