+++ shared/models/users/user-flag.model.ts
@@ -1,4 +1,4 @@
-export enum UserAdminFlag {
+export const enum UserAdminFlag {
   NONE = 0,
   BYPASS_VIDEO_AUTO_BLACKLIST = 1 << 0
 }
