+++ shared/models/videos/playlist/video-playlist-privacy.model.ts
@@ -1,4 +1,4 @@
-export enum VideoPlaylistPrivacy {
+export const enum VideoPlaylistPrivacy {
   PUBLIC = 1,
   UNLISTED = 2,
   PRIVATE = 3
