+++ shared/models/videos/playlist/video-playlist-type.model.ts
@@ -1,4 +1,4 @@
-export enum VideoPlaylistType {
+export const enum VideoPlaylistType {
   REGULAR = 1,
   WATCH_LATER = 2
 }
