+++ shared/models/videos/thumbnail.type.ts
@@ -1,4 +1,4 @@
-export enum ThumbnailType {
+export const enum ThumbnailType {
   MINIATURE = 1,
   PREVIEW = 2
 }
