+++ shared/models/videos/video-privacy.enum.ts
@@ -1,4 +1,4 @@
-export enum VideoPrivacy {
+export const enum VideoPrivacy {
   PUBLIC = 1,
   UNLISTED = 2,
   PRIVATE = 3,
