+++ shared/models/videos/video-state.enum.ts
@@ -1,4 +1,4 @@
-export enum VideoState {
+export const enum VideoState {
   PUBLISHED = 1,
   TO_TRANSCODE = 2,
   TO_IMPORT = 3
