+++ shared/models/videos/video-streaming-playlist.type.ts
@@ -1,3 +1,3 @@
-export enum VideoStreamingPlaylistType {
+export const enum VideoStreamingPlaylistType {
   HLS = 1
 }
